EIP: restrict EIP controller only to networks that have role primary
..and the EIP controller owns the aforementioned network.

+++ go-controller/pkg/ovn/base_network_controller_egressip.go
@@ -188,7 +188,18 @@ func (bnc *BaseNetworkController) reconcileEgressIP(old, new *egressipv1.EgressI
 			if err != nil {
 				return err
 			}
+			var role string
 			for _, namespace := range namespaces {
+				// determine if this network configures EIP for this namespace
+				role, err = bnc.GetNetworkRoleForNamespace(namespace)
+				if err != nil {
+					return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+				}
+				if role != types.NetworkRolePrimary {
+					// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+					// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+					continue
+				}
 				namespaceLabels := labels.Set(namespace.Labels)
 				if !newNamespaceSelector.Matches(namespaceLabels) && oldNamespaceSelector.Matches(namespaceLabels) {
 					if err := bnc.deleteNamespaceEgressIPAssignment(oldEIP.Name, oldEIP.Status.Items, namespace, oldEIP.Spec.PodSelector); err != nil {
@@ -210,7 +221,18 @@ func (bnc *BaseNetworkController) reconcileEgressIP(old, new *egressipv1.EgressI
 			if err != nil {
 				return err
 			}
+			var role string
 			for _, namespace := range namespaces {
+				// determine if this network configures EIP for this namespace
+				role, err = bnc.GetNetworkRoleForNamespace(namespace)
+				if err != nil {
+					return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+				}
+				if role != types.NetworkRolePrimary {
+					// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+					// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+					continue
+				}
 				pods, err := bnc.watchFactory.GetPods(namespace.Name)
 				if err != nil {
 					return err
@@ -241,7 +263,18 @@ func (bnc *BaseNetworkController) reconcileEgressIP(old, new *egressipv1.EgressI
 			if err != nil {
 				return err
 			}
+			var role string
 			for _, namespace := range namespaces {
+				// determine if this network configures EIP for this namespace
+				role, err = bnc.GetNetworkRoleForNamespace(namespace)
+				if err != nil {
+					return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+				}
+				if role != types.NetworkRolePrimary {
+					// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+					// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+					continue
+				}
 				namespaceLabels := labels.Set(namespace.Labels)
 				// If the namespace does not match anymore then there's no
 				// reason to look at the pod selector.
@@ -311,7 +344,16 @@ func (bnc *BaseNetworkController) reconcileEgressIPNamespace(old, new *v1.Namesp
 	if new != nil {
 		newNamespace = new
 	}
-
+	// determine if this network supports EIP for this namespace
+	role, err := bnc.GetNetworkRoleForNamespace(getFirstNonNilNamespace(newNamespace, oldNamespace))
+	if err != nil {
+		return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+	}
+	if role != types.NetworkRolePrimary {
+		// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+		// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+		return nil
+	}
 	// If the labels have not changed, then there's no change that we care
 	// about: return.
 	oldLabels := labels.Set(oldNamespace.Labels)
@@ -375,7 +417,16 @@ func (bnc *BaseNetworkController) reconcileEgressIPPod(old, new *v1.Pod) (err er
 			return err
 		}
 	}
-
+	// determine if this network supports EIP for this namespace
+	role, err := bnc.GetNetworkRoleForNamespace(namespace)
+	if err != nil {
+		return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+	}
+	if role != types.NetworkRolePrimary {
+		// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+		// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+		return nil
+	}
 	newPodLabels := labels.Set(newPod.Labels)
 	oldPodLabels := labels.Set(oldPod.Labels)
 
@@ -473,7 +524,18 @@ func (bnc *BaseNetworkController) addEgressIPAssignments(name string, statusAssi
 	if err != nil {
 		return err
 	}
+	var role string
 	for _, namespace := range namespaces {
+		// determine if this network configures EIP for this namespace
+		role, err = bnc.GetNetworkRoleForNamespace(namespace)
+		if err != nil {
+			return fmt.Errorf("failed to configure EgressIP for Namespace because unable to retrieve network role: %w", err)
+		}
+		if role != types.NetworkRolePrimary {
+			// nothing to clean up as this network is not the primary network of this namespace. We don't have to be concerned
+			// if network role changes from primary, because the network will be deleted and recreated by net-con-mgr.
+			return nil
+		}
 		if err := bnc.addNamespaceEgressIPAssignments(name, statusAssignments, namespace, podSelector); err != nil {
 			return err
 		}
@@ -1194,6 +1256,16 @@ func (bnc *BaseNetworkController) generateCacheForEgressIP() (map[string]egressI
 			continue
 		}
 		for _, namespace := range namespaces {
+			// determine if this network configures EIP for this namespace
+			role, err := bnc.GetNetworkRoleForNamespace(namespace)
+			if err != nil {
+				klog.Errorf("Failed to generate EgressIP cache for Namespace %s because unable to retrieve network role: %w", namespace.Name, err)
+				continue
+			}
+			// skip config if network isn't primary
+			if role != types.NetworkRolePrimary {
+				continue
+			}
 			pods, err := bnc.watchFactory.GetPodsBySelector(namespace.Name, egressIP.Spec.PodSelector)
 			if err != nil {
 				klog.Errorf("Error building egress IP sync cache, cannot retrieve pods for namespace: %s and egress IP: %s, err: %v", namespace.Name, egressIP.Name, err)
@@ -2306,4 +2378,31 @@ func getPodKey(pod *kapi.Pod) string {
 func getPodNamespaceAndNameFromKey(podKey string) (string, string) {
 	parts := strings.Split(podKey, "_")
 	return parts[0], parts[1]
+}
+
+// getFirstNonNilNamespace must return a non-nil pointer to a Namespace otherwise panic
+func getFirstNonNilNamespace(n1, n2 *v1.Namespace) *v1.Namespace {
+	if n1 != nil {
+		return n1
+	}
+	if n2 != nil {
+		return n2
+	}
+	panic("getNonNilNamespace(): received two nil Namespace pointers")
+}
+
+func addPktMarkToLRPOptions(options map[string]string, mark string) {
+	options["pkt_mark"] = mark
+}
+
+func getEgressIPPktMark(eipName string, annotations map[string]string) util.EgressIPMark {
+	var err error
+	var mark util.EgressIPMark
+	if util.IsNetworkSegmentationSupportEnabled() && util.IsEgressIPMarkSet(annotations) {
+		mark, err = util.ParseEgressIPMark(annotations)
+		if err != nil {
+			klog.Errorf("Failed to get EgressIP %s packet mark from annotations: %v", eipName, err)
+		}
+	}
+	return mark
 }
